Reorganize package structure for improved modularity and clarity
Refactors the codebase by moving classes into more semantically appropriate packages, enhancing modularity and project organization.
