+++ app/src/main/java/de/uol/neuropsy/senda/sensor/AudioBridge.kt
@@ -1,4 +1,4 @@
-package de.uol.neuropsy.senda
+package de.uol.neuropsy.senda.sensor
 
 import android.content.Context
 import android.media.AudioFormat
