+++ app/src/main/res/layout-land/activity_main.xml
@@ -10,7 +10,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="de.uol.neuropsy.senda.MainActivity"
+    tools:context="de.uol.neuropsy.senda.ui.main.MainActivity"
     tools:layout_editor_absoluteY="25dp">
 
     <LinearLayout
