+++ app/src/main/res/layout/activity_main.xml
@@ -9,7 +9,7 @@
                                                    android:paddingTop="@dimen/activity_vertical_margin"
                                                    android:paddingRight="@dimen/activity_horizontal_margin"
                                                    android:paddingBottom="@dimen/activity_vertical_margin"
-                                                   tools:context="de.uol.neuropsy.senda.MainActivity"
+                                                   tools:context="de.uol.neuropsy.senda.ui.main.MainActivity"
                                                    tools:layout_editor_absoluteY="25dp">
 
     <ImageView
